+++ web/src/screens/settings/index.ts
@@ -3,13 +3,13 @@
  * SPDX-License-Identifier: GPL-2.0-or-later
  */
 
-export { default as APISettings } from "./Api";
-export { default as ApplicationSettings } from "./Application";
-export { default as DownloadClientSettings } from "./DownloadClient";
-export { default as FeedSettings } from "./Feed";
-export { default as IndexerSettings } from "./Indexer";
-export { default as IrcSettings } from "./Irc";
-export { default as LogSettings } from "./Logs";
-export { default as NotificationSettings } from "./Notifications";
-export { default as ReleaseSettings } from "./Releases";
+export { default as Api } from "./Api";
+export { default as Application } from "./Application";
+export { default as DownloadClient } from "./DownloadClient";
+export { default as Feed } from "./Feed";
+export { default as Indexer } from "./Indexer";
+export { default as Irc } from "./Irc";
+export { default as Logs } from "./Logs";
+export { default as Notification } from "./Notifications";
+export { default as Release } from "./Releases";
 export { default as RegexPlayground } from "./RegexPlayground";
